Moved event, error and task handlers into one folder
