+++ files/reactionroles.json
@@ -1,1 +1,16 @@
-[]
+[
+    {
+        "guild_id": 953054451999072276,
+        "name": "Python \ud83d\udc0d",
+        "role_id": 953074155681624074,
+        "emoji": "\ud83c\udfd6\ufe0f",
+        "msg_id": 955562843896447086
+    },
+    {
+        "guild_id": 953054451999072276,
+        "name": "JavaScript \ud83d\udcdc",
+        "role_id": 953260218421243915,
+        "emoji": "\ud83c\udde8\ud83c\udde6",
+        "msg_id": 955567503315177552
+    }
+]
