+++ start.py
@@ -6,13 +6,10 @@ This is where the bot is configured and launched.
 
 import nextcord
 import os
-import nextcord.ext.commands as commands
-import nextcord.ext.tasks as tasks
 import itertools
 import dotenv
 
-
-# |---------- CONFIG ----------|
+from nextcord.ext import commands, tasks
 
 
 dotenv.load_dotenv('files/.env')
@@ -22,7 +19,7 @@ intents.members = True
 
 prefix = "~"
 colour = 0x486572
-status = itertools.cycle(['~help', '~ai', '~play'])
+status = itertools.cycle(['~help', '~play'])
 client = commands.Bot(prefix, intents=intents, case_insensitive=True) # Help command = none so we can override it
 
 # List of filepaths in the Python import format
@@ -37,46 +34,19 @@ cogs = [
 
 for cog in cogs:
     client.load_extension(cog)
-    print(f'Loaded "{cog}" successfully')
+    print(f'[COG] Loaded "{cog}" successfully')
 
+# Doing the same for handlers
+# This is separate to improve readability
+handlers = [
+    'handlers.error_handler',
+    'handlers.event_handler',
+    'handlers.task_handler'
+]
 
-# def load(module):
-#     if module.endswith('.py') and module not in ('__init__.py', 'helpers.py'):
-#         client.load_extension(f'modules.{module}'.replace('.py', ''))
-
-
-# for module in os.listdir(PATH):
-#     if module == HELP:
-#         load(f'{HELP}.help_cog.py')
-
-#     load(module)
-
-
-# |----- BACKGROUND TASKS -----|
-
-
-@tasks.loop(seconds=30)
-async def change_presence():
-    """
-    Changes the bot's presence every 30 seconds.
-    """
-    activity = next(status)
-    await client.change_presence(activity=nextcord.Game(activity))
-
-
-# |---------- EVENTS ----------|
-
-
-@client.event
-async def on_ready():
-    """
-    Executes when the bot has loaded.
-    """
-    print(f'Loaded {client.user.name} successfully.')
-    await change_presence.start()
-
-
-# |---------- LAUNCH ----------|
+for handler in handlers:
+    client.load_extension(handler)
+    print(f'[HANDLER] Loaded "{handler}" successfully')
 
 
 TOKEN = os.getenv("TOKEN")
